Filter displayed devices by the membership of the user, according to his user and device groups. An admin user group will be automatically created. Any user is admin, if this admin group is empty for maintenence reasons, if not any user belong to the admin group is admin. All physically present devices are shown to admins, otherwise only devices are shown to the user according to his groups. Only allow users to use devices according to his groups and limitations.

+++ res/app/settings/groups/device-groups/device-groups-controller.js
@@ -24,15 +24,21 @@ module.exports = function DeviceGroupsCtrl($scope, DeviceGroupsService) {
   }
 
   $scope.removeDeviceGroup = function(group) {
-    DeviceGroupsService.removeDeviceGroup(group)
+    if (confirm('Do you want to remove the device group?')) {
+      DeviceGroupsService.removeDeviceGroup(group)
+    }
   }
 
   $scope.removeDeviceFromGroup = function(group, device) {
-    DeviceGroupsService.removeDeviceFromGroup(group, device)
+    if (confirm('Do you want to remove the device from the group?')) {
+      DeviceGroupsService.removeDeviceFromGroup(group, device)
+    }
   }
 
   $scope.removeUserGroupFromGroup = function(group, userGroupTitle) {
-    DeviceGroupsService.removeAssociatedUserGroupFromGroup(group, userGroupTitle)
+    if (confirm('Do you want to remove the user group from the device group?')) {
+      DeviceGroupsService.removeAssociatedUserGroupFromGroup(group, userGroupTitle)
+    }
   }
 
   $scope.$on('group.device.group.updated', updateDeviceGroups)
