+++ res/app/settings/groups/user-groups/user-groups-controller.js
@@ -1,6 +1,7 @@
 module.exports = function UserGroupsCtrl($scope, UserGroupsService) {
 
   $scope.userGroups = []
+  $scope.maxSafeInteger = Number.MAX_SAFE_INTEGER
 
   function updateUserGroups() {
     UserGroupsService.getUserGroups()
@@ -24,11 +25,15 @@ module.exports = function UserGroupsCtrl($scope, UserGroupsService) {
   }
 
   $scope.removeUserGroup = function(group) {
-    UserGroupsService.removeUserGroup(group)
+    if (confirm('Do you want to remove the user group?')) {
+      UserGroupsService.removeUserGroup(group)
+    }
   }
 
   $scope.removeUserFromGroup = function(group, user) {
-    UserGroupsService.removeUserFromGroup(group, user)
+    if (confirm('Do you want to remove the user from the group?')) {
+      UserGroupsService.removeUserFromGroup(group, user)
+    }
   }
 
   $scope.$on('group.user.group.updated', updateUserGroups)
